Using MOOLE_400_STABLE version of externallib.php
We have custom changes in that have not made it upstream. They live in
commit 8c870fbb373037e4bdc54406cdaa879f164b1317 by Seth Yoder.

When upgrading to 4.1, I merged the MOODLE_401_STABLE branch into our
aprende branch, under the assumption that it was the version for 4.1.
This was incorrect -- while most of the code between MOODLE_401_STABLE
and MOODLE_400_STABLE is the same, the 401 branch had changes that were
made for Moodle 4.2.

The namespace for external_api changes in 4.2, this breaks the
mod_questionnaire external function and the documentation / webservice
modification page.

I fixed this by just copying the externallib.php file from MOODLE_400